Hide some functions from vecmeth public interface
