+++ src/vecmath/normalize.rs
@@ -9,6 +9,7 @@ where
     Self: Sized + Div<T, Output = Self> + Copy + Clone,
     T: Mul<T, Output = T> + Add<T, Output = T> + Sqrt,
 {
+    // TODO I think we can take a reference here instead.
     fn normalize(self) -> Self {
         debug_assert!(!self.has_nan());
         self / self.length()
